feat: move server monitor to dedicated module
Extract the server_monitor function from the main binary into a new
monitors/server.rs module to improve code organization and separation
of concerns. Update imports and module declarations accordingly.

This refactoring makes the codebase more modular and maintainable,
preparing for future enhancements such as running monitors in separate
tasks or threads.
