+++ src/bin/hub.rs
@@ -1,13 +1,10 @@
-use std::time::Duration;
-
 use clap::Parser;
 use server_monitoring::{
-    ServerMetrics,
-    config::{Config, ServerConfig, read_config_file},
-    resource_monitor::ResourceMonitor,
+    config::{Config, read_config_file},
+    monitors::server::server_monitor,
 };
 use tokio::{join, spawn};
-use tracing::{debug, error, instrument, level_filters::LevelFilter, trace};
+use tracing::{error, level_filters::LevelFilter, trace};
 use tracing_subscriber::{filter, layer::SubscriberExt, util::SubscriberInitExt};
 
 #[derive(Debug, Clone, Parser)]
@@ -62,66 +59,5 @@ async fn dispatch_servers(config: &Config) {
         if let Err(e) = handler.await {
             error!("{e}");
         }
-    }
-}
-
-#[instrument(skip_all)]
-async fn server_monitor(config: ServerConfig) {
-    let ServerConfig {
-        ip,
-        port,
-        interval,
-        token,
-        display,
-        ..
-    } = config.clone();
-    let display_name = display.unwrap_or(String::from("unknown"));
-    debug!("starting server monitor for {display_name} ({ip}:{port}) with interval {interval}");
-
-    let url = format!("http://{ip}:{port}/metrics");
-
-    let mut monitor = ResourceMonitor::new(config);
-
-    loop {
-        tokio::time::sleep(Duration::from_secs(interval as u64)).await;
-
-        trace!("{url}: requesting metrics");
-
-        let client = reqwest::Client::new();
-        let request = client.get(&url).header(
-            "X-MONITORING-SECRET",
-            token.as_ref().unwrap_or(&String::new()),
-        );
-
-        let response = request.send().await;
-        let body = match response {
-            Ok(body) => body,
-            Err(e) => {
-                error!("{url}: error during request: {e}");
-                continue;
-            }
-        };
-
-        let body = match body.text().await {
-            Ok(body) => body,
-            Err(e) => {
-                error!("{url}: error during decode: {e}");
-                continue;
-            }
-        };
-
-        let metrics = match serde_json::from_str::<ServerMetrics>(&body) {
-            Ok(metrics) => metrics,
-            Err(e) => {
-                error!("{url}: error while trying to parse the metrics: {e}");
-                continue;
-            }
-        };
-
-        trace!("{url}: received metrics");
-
-        // trace!("received server metrics for {url}: {metrics:?}");
-
-        monitor.update(&metrics).await;
     }
 }
